fixy warnings - I hope

+++ src/main/java/erebus/client/render/entity/RenderPunchroom.java
@@ -24,7 +24,7 @@ public class RenderPunchroom extends RenderLiving {
 	protected void preRenderCallback(EntityLivingBase entityliving, float partialTickTime) {
 		EntityPunchroom punchroom = (EntityPunchroom) entityliving;
         int i = 1;
-        float f1 = (punchroom.prevSquishFactor + (punchroom.squishFactor - punchroom.prevSquishFactor) * partialTickTime) / ((float)i * 0.5F + 1.0F);
+        float f1 = (punchroom.prevSquishFactor + (punchroom.squishFactor - punchroom.prevSquishFactor) * partialTickTime) / (i * 0.5F + 1.0F);
         float f2 = 1.0F / (f1 + 1.0F);
         float f3 = (float)i;
         GL11.glScalef(f2 * f3, 1.0F / f2 * f3, f2 * f3);
